Configure one-step registration workflow
To enable a simple registration workflow (users sign up and their
accounts are immediately active and logged in) I here include
the django-registration simple backend URLS and add a template
for the registration form.

This URLconf will also configure the appropriate URLs for the
rest of the built-in django.contrib.auth views (log in, log out,
password reset, etc.). I have not yet created templates for these
auth views.

If made available to the public we'd want better control over user
account registration, and this could easily be achieved by
switching to django-registration's HMAC activation workflow, which
implements a two-step process: a user signs up, then is emailed an
activation link and must click it to activate the account.